rename enc to encrypt
