+++ code/logic/fossil/cryptic/encrypt.h
