+++ code/logic/fossil/cryptic/framework.h
@@ -27,6 +27,7 @@
 
 #include "hash.h"
 #include "auth.h"
-#include "enc.h"
+#include "encrypt.h"
+#include "dencrypt.h"
 
 #endif /* FOSSIL_CRYPTIC_FRAMEWORK_H */
